requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
frozendict==2.4.6
html5lib==1.1
idna==3.10
isodate==0.7.2
lxml==5.3.0
msal==1.31.1
msal-extensions==1.2.0
multitasking==0.0.11
numpy==2.1.3
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
portalocker==2.10.1
py4j==0.10.9.7
pycparser==2.22
PyJWT==2.10.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.2.50
